19-010 Handling Http States & Feedback with Redux-A
added a Notification component to display application's status
Application's status is monitered by ui-slice
Once any status need to be paged
we can dispatch Notification to the ui-slice.
In the App component, we can display this Notifition to user

One issue need to be fixed: as the useEffect depends on "cart" state, when the Application is launching, the STORE will init the Cart state with initial value. this initiation will trigger the HTTP Request.  It is bad.