19-013 Finalizing the Fetching Logic - A
after fetching the data, the Application will fire an un-expected Http request and Send data to Firebase.

The reason is once we fetched the data. the State was updated. which will trigger "sendCartData"

For fixing this issue, add a "changed" property  to make the trick

The FetchCartData use "replaceCart reducer which will not trigger "changed" to TRUE
so that FetchCartData will not fire SendCartData Http request again.